Change data types of context, phrase and clue sprint

+++ database/migrations/2019_07_16_105649_create_phrase_sprint.php
@@ -16,9 +16,9 @@ class CreatePhraseSprint extends Migration
         Schema::connection('pgsql')->create('phrase_sprint', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('topic_id');
-            $table->unsignedInteger('phrase_id');
-            $table->unsignedInteger('solution_context_id');
-            $table->unsignedInteger('wrong_context_id');
+            $table->unsignedBigInteger('phrase_id');
+            $table->unsignedBigInteger('solution_context_id');
+            $table->unsignedBigInteger('wrong_context_id');
             $table->integer('bucket');
             $table->foreign('topic_id')->references('id')->on('context_topics')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
